requirements.txt
requests>=2.32.0
python-dotenv>=1.0.1
psutil>=5.9.8
Pillow>=10.4.0
pystray>=0.19.5
